Extract, convert parseWordList to TS
